Upcoming Payment details is working

+++ composeApp/src/commonMain/kotlin/dev/propoc/moneysaver/core/presentation/TextUtils.kt
@@ -1,4 +1,4 @@
 package dev.propoc.moneysaver.core.presentation
 
-fun Int.moneyFormat() = "$$this"
+fun Double.moneyFormat() = "$$this"
